\name{data1}

\alias{data1}

\docType{data}

\title{A simulated dataset}

\description{
A data frame with simulated genotype and phenotype
}

\usage{data(data1)}

\format{
  A data frame with 100 observations on the following 21 variables.
  \describe{The first column is the phenotypic data, and other 20 columns are the genotypic data.}
}

\details{
The phenotypic data was generated using the first four predictive variables(x1-x4) by interacting with x1 and x2, x3 and x4, resepectively. 
}

\examples{
data(data1)
}

\keyword{datasets}
