\name{data2}

\alias{data2}

\docType{data}

\title{A simulated dataset}

\description{
A data frame with simulated genotype and phenotype
}

\usage{data(data2)}

\format{
  A data frame with 100 observations on the following 16 variables.
  \describe{The first column is the phenotypic data, and other 15 columns are the genotypic data.}
}

\details{
The phenotypic data was generated using the first six predictive variables(x1-x6) by interacting with x1 x2 and x3, x4 x5 and x6, resepectively. 
}

\examples{
data(data2)
}

\keyword{datasets}
