\name{MDRforward-package}
\alias{MDRforward-package}
\alias{MDRforward}
\docType{package}
\title{
A MDR based forward selection method 
}
\description{
A MDR (multi-factor dimensionality reduction) based forward selection method for genetic association mapping.
}

\details{
\tabular{ll}{
Package: \tab MDRforward\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-04\cr
License: GPL(>= 2)\cr
}

}
\author{
Yi Xu and Jixiang Wu\cr
Maintainer: Yi Xu <statxy@gmail.com>
}

\references{

Marylyn D. Ritchie, Lance W. Hahn, Nady Roodi, L. Renee Bailey, William D. Dupont, Fritz F. Parl, and Jason H. Moore (2001) Multifactor-dimensionality reduction reveals high-order interactions among estrogen-metabolism genes in sporadic breast cancer. American Journal of Human Genetics 69: 138-147.

Yi Xu, Jixiang Wu, Detecting epistatic effects among SNP markers associated with three barley traits by a MDR based forward selection method (unpublished).

Yi Xu, Jixiang Wu, Detecting higher-order interactions of SNP markers associated with three barley agronomic traits (unpublished).

}

\keyword{ package }

